==> fit_model/5-fit_model.R <==
rm(list=ls())

setwd("/Users/DGravel/Documents/Projects_On_Going/Maple_migration/talk_Ouranos/fit_model")
source("4-init_params_newData.R")
source("3-transition_model.R")

# Maximum likelihood estimation
library(GenSA)

#test
model(params, data)

estim.pars = GenSA(par = params, fn = model, lower = par_lo, upper= par_hi, control = list(verbose =TRUE, max.time = 172800, smooth=FALSE), dat = data)

#save(estim.pars, file="../estimated_params/GenSA_test.rdata")
names(estim.pars$par) = unlist(lapply(names(params), function(x){strsplit(x, split = ".", fixed= TRUE)[[1]][[1]]}))
write.table(estim.pars$par,file="GenSA_test_subTemp.txt")
